chore: clean up the code
